MVP
Proficient E-Commerce and Inventory Control System

🚧 Project Status: Work in progress (WIP)

The current project is an extensive inventory control and e-commerce system that is built based on PHP and uses a combination of the repository/service pattern and MVC standards for implementation.

Current Milestone: MVP Deployment

The MVP of the system is now stable, allowing the core system to provide the following important workflows for public deployment:

User Login and User Profile Management

Product Browse and Product Filter and Product Display

Shopping Cart Management

Order Creation and Payment Processing

Admin Panel to administer all system entities (e.g., Products, Users, Stocks, Categories, etc.)

📝 Next Steps and Technical Issues (To Be Tackled Post-Exams)

Though the Minimum Viable Product is now ready to deploy, a significant amount of technical debt remains that must be corrected for the long-term viability and security of the product.

1. API and Back End Refactor

The focus for this milestone is to create and implement a more sustainable and maintainable core of PHP logic to the system. This will also address issues related to security and lack of validation.

1a. Validation Refactor - This will enable the consolidation and unifying of the validation processes for all validator types (admin/validators) into a single location. The intention is to maintain uniformity within the application and reduce security risks. There are numerous Validator files that can have common validation techniques.

1b. User API Endpoint - This will separate and partition the current User API Endpoints; as such, it will allow developers to apply the Single Responsibility principle to each of these APIs (e.g., to distinguish between authentication-related User APIs and profile-related User APIs).